test: cli refactor
