+++ test/cli/compress-option.test.js
@@ -1,0 +1,17 @@
+'use strict';
+
+const { testBin } = require('../helpers/test-bin');
+
+describe('compress option', () => {
+  it('--compress', async () => {
+    const { exitCode } = await testBin('--compress');
+
+    expect(exitCode).toEqual(0);
+  });
+
+  it('--no-compress', async () => {
+    const { exitCode } = await testBin('--no-compress');
+
+    expect(exitCode).toEqual(0);
+  });
+});
