+++ test/cli/liveReload-option.test.js
@@ -1,0 +1,17 @@
+'use strict';
+
+const { testBin } = require('../helpers/test-bin');
+
+describe('liveReload option', () => {
+  it('--live-reload', async () => {
+    const { exitCode } = await testBin(['--live-reload']);
+
+    expect(exitCode).toEqual(0);
+  });
+
+  it('--no-live-reload', async () => {
+    const { exitCode } = await testBin(['--no-live-reload']);
+
+    expect(exitCode).toEqual(0);
+  });
+});
